Move getNetworkIdentity to network class
