+++ packages/cli/src/cmds/beacon/handler.ts
@@ -83,10 +83,12 @@ export async function beaconHandler(args: BeaconArgs & GlobalArgs): Promise<void
 
     onGracefulShutdown(async () => {
       if (args.persistNetworkIdentity) {
-        const enr = await node.network.getEnr().catch((e) => logger.warn("Unable to persist enr", {}, e));
-        if (enr) {
+        try {
+          const networkIdentity = await node.network.getNetworkIdentity();
           const enrPath = path.join(beaconPaths.beaconDir, "enr");
-          writeFile600Perm(enrPath, enr.encodeTxt());
+          writeFile600Perm(enrPath, networkIdentity.enr);
+        } catch (e) {
+          logger.warn("Unable to persist enr", {}, e as Error);
         }
       }
       abortController.abort();
